requirements.txt
# Core dependencies (optimized)
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.15.0
yfinance>=0.2.18
textblob>=0.17.1
scikit-learn>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0

# File processing dependencies
pdfplumber>=0.9.0
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.0